backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
boto3==1.35.81
botocore==1.35.81
httpx==0.28.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
langgraph==0.1.57
langchain==0.1.20
langchain-openai==0.1.15
langchain-community==0.1.21
langchain-upstage==0.0.15
pytest==7.4.4
pytest-asyncio==0.23.2
